backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
python-multipart # For form data

# Google Cloud Libraries
google-cloud-speech
google-cloud-texttospeech
google-cloud-translate
google-maps-routing
google-generativeai
googlemaps

# Audio/Numeric
pydub
noisereduce
scipy # Dependency for noisereduce
numpy # Dependency for noisereduce

# Computer Vision/ML for Drowsiness
opencv-python
mediapipe
ultralytics
torch
torchvision
pyyaml